+++ layouts/DashboardLayout.tsx
@@ -2,7 +2,7 @@ import { useState, PropsWithChildren } from 'react'
 import { useRouter } from 'next/router'
 import { Box, Container, Grid, NavLink, Navbar, Group, Paper } from '@mantine/core';
 import { usePathname } from 'next/navigation';
-import { dashboardElements } from '../library/components/Dashboard/DashboardHelper'
+import { dashboardElements } from '../library/pages/Dashboard/DashboardHelper'
 import { IconUser, IconBookmarks, IconActivity, IconChevronRight, IconLock } from '@tabler/icons-react';
