--- library/components/ColorSchemeToggle/ColorSchemeToggle.tsx
@@ -1,26 +1,0 @@
-import { ActionIcon, Group, useMantineColorScheme } from '@mantine/core';
-import { IconSun, IconMoonStars } from '@tabler/icons-react';
-
-export function ColorSchemeToggle() {
-  const { colorScheme, toggleColorScheme } = useMantineColorScheme();
-
-  return (
-    <Group position="center">
-      <ActionIcon
-        onClick={() => toggleColorScheme()}
-        size="lg"
-        sx={(theme) => ({
-          backgroundColor:
-            theme.colorScheme === 'dark' ? theme.colors.dark[6] : theme.colors.gray[0],
-          color: theme.colorScheme === 'dark' ? theme.colors.yellow[4] : theme.colors['raga-red'][8],
-        })}
-      >
-        {colorScheme === 'dark' ? (
-          <IconSun size={20} stroke={1.5} />
-        ) : (
-          <IconMoonStars size={20} stroke={1.5} />
-        )}
-      </ActionIcon>
-    </Group>
-  );
-}
