+++ library/components/raga_components/RagaModal/RagaModal.tsx
@@ -11,7 +11,7 @@ import {
 } from '@mantine/core';
 import { IconPlayerPlay, IconHeart, IconBookmark, IconShare } from '@tabler/icons-react'
 import { RagaCardProps } from '../RagaCard/RagaCard'
-import { SwaraGradient } from '../../SwaraGradient/SwaraGradient';
+import { SwaraGradient } from '../SwaraGradient/SwaraGradient';
 import { Swara, swaraNoteMap } from '../../../helpers/SwaraHelpers'
 import Tone from '../../Tone'
